conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
brotlipy=0.7.0=py36h2bbff1b_1003
ca-certificates=2022.4.26=haa95532_0
certifi=2021.5.30=py36haa95532_0
cffi=1.14.6=py36h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=35.0.0=py36h71e12ea_0
cudatoolkit=11.1.1=heb2d755_10
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh787bdff_2
freetype=2.8.1=ha63716d_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
intel-openmp=2022.1.0=h57928b3_3787
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.1=py36hd77b12b_0
libpng=1.6.37=h1d00b33_2
libtiff=4.0.9=hf1753bf_1
libuv=1.43.0=h8ffe710_0
lxml=4.9.0=pypi_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=2.2.2=py36h153e9ff_1
mkl=2020.2=256
mkl-include=2022.1.0=h6a75c08_874
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.3.0=py36h46781fe_0
mkl_random=1.1.1=py36h47e9c7a_0
msys2-conda-epoch=20160418=1
multitasking=0.0.10=pypi_0
ninja=1.11.0=h2d74725_0
numpy=1.19.2=py36hadc3359_0
numpy-base=1.19.2=py36ha3acd2a_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1o=h2bbff1b_0
pandas=1.1.5=py36hd77b12b_0
pillow=5.2.0=py36_0
pip=21.2.2=py36haa95532_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h6538335_2
pysocks=1.7.1=py36haa95532_0
python=3.6.13=h3758d61_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.6=2_cp36m
pytorch=1.8.0=py3.6_cuda11.1_cudnn8_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
qt=5.9.7=vc14h73c81de_0
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=0.24.2=py36hf11a4ad_1
scipy=1.5.2=py36h9439919_0
selenium=3.141.0=py36h2bbff1b_1000
setuptools=58.0.4=py36haa95532_0
sip=4.19.8=py36h6538335_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.3=h2bbff1b_0
tbb=2021.5.0=h2d74725_1
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h8ffe710_0
torchaudio=0.8.0=py36
torchvision=0.9.0=py36_cu111
tornado=6.1=py36h2bbff1b_0
typing_extensions=4.1.1=pyha770c72_0
urllib3=1.26.8=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py36haa95532_0
wincertstore=0.2=py36h7fe50ca_0
xlrd=2.0.1=pyhd3eb1b0_0
yahoo-finance-api2=0.0.12=pypi_0
yfinance=0.1.70=pypi_0
zlib=1.2.11=h8d14728_1011
